Minor fixes.

+++ libdf3d/io/FileDataSource.h
@@ -24,7 +24,7 @@ public:
     virtual size_t getSizeInBytes() = 0;
 
     virtual size_t tell() = 0;
-    virtual bool seek(size_t offset, std::ios_base::seekdir origin) = 0;
+    virtual bool seek(int32_t offset, std::ios_base::seekdir origin) = 0;
 
     template<typename T>
     bool getAsObjects(T *output, size_t numElements)
